Fixed `PlacementIndications`. Breaking change!
