+++ bsb/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "4.0.0a6"
+__version__ = "4.0.0a7"
 
 import functools
