+++ bsb/cell_types.py
@@ -11,43 +11,6 @@ from .exceptions import *
 import abc
 
 
-@config.dynamic(
-    attr_name="selector",
-    auto_classmap=True,
-    required=False,
-    default="by_name",
-)
-class MorphologySelector(abc.ABC):
-    @abc.abstractmethod
-    def validate(self, all_morphos):
-        pass
-
-    @abc.abstractmethod
-    def pick(self, morphology):
-        pass
-
-
-@config.node
-class NameSelector(MorphologySelector, classmap_entry="by_name"):
-    names = config.list(type=str)
-
-    def validate(self, all_morphos):
-        missing = set(self.names) - {m.get_meta()["name"] for m in all_morphos}
-        if missing:
-            raise MissingMorphologyError(
-                f"Morphology repository misses the following morphologies required by {self._config_parent._config_parent.get_node_name()}: {', '.join(missing)}"
-            )
-
-    def pick(self, morphology):
-        return morphology.get_meta()["name"] in self.names
-
-
-@config.node
-class Representation(PlacementIndications):
-    geometrical = config.dict(type=types.any())
-    morphological = config.list(type=MorphologySelector)
-
-
 @config.node
 class Plotting:
     display_name = config.attr()
@@ -67,7 +30,7 @@ class CellType:
 
     name = config.attr(key=True)
     spatial = config.attr(
-        type=Representation, required=_not_an_entity, default={"radius": None}
+        type=PlacementIndications, required=_not_an_entity, default={"radius": None}
     )
     plotting = config.attr(type=Plotting)
     relay = config.attr(type=bool, default=False)
